Swap bit-endianness of value in note commitment.
