+++ src/lib.rs
@@ -15,3 +15,4 @@ pub mod circuit;
 pub mod pedersen_hash;
 pub mod primitives;
 pub mod constants;
+mod util;
